--- modules/certbund/defaults.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-
-NAME = "CERT Bund (BSI)"
-CHANNELS = (
-    "newsfeed",
-)
-URL = "https://www.bsi.bund.de/SiteGlobals/Functions/RSSFeed/RSSNewsfeed/RSSNewsfeed_CSW.xml"
-ENTRIES = 10
-TRANSLATION = True
