+++ modules/certde/defaults.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+
+NAME = "CERT Bundesrepublik Deutschland"
+CHANNELS = (
+    "newsfeed",
+)
+URLS = (
+    "https://wid.cert-bund.de/content/public/securityAdvisory/rss",
+    "https://www.bsi.bund.de/SiteGlobals/Functions/RSSFeed/RSSNewsfeed/RSSNewsfeed_CSW.xml"
+)
+ENTRIES = 10
+TRANSLATION = True
